requirements.txt
ultralytics
opencv-python
numpy
#vpi ~=2.3
pandas
PyYAML
kafka-python
# deepface
